fix: reaction display, and component moved to icons directory
