+++ src/components/icons/LikeFilled.svelte
